Renamed proejct to RecolorShaderUnity
