+++ .github/variables/projectconfig.env
@@ -1,3 +1,3 @@
-PROJECT_NAME=TemplateUnityPackage
+PROJECT_NAME=RecolorShaderUnity
 UNITY_VERSION=2021.3.11f1
-PACKAGE_PATH=./Packages/com.nickmaltbie.templateunitypackage
+PACKAGE_PATH=./Packages/com.nickmaltbie.recolorshaderunity
