+++ Assets/Editor/AssetGUIDRegenerator.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6f16f2d106992a34b8db7e3a2691be98
+guid: 3209990e5fc3b824aa1d774ac734c9e3
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
