+++ Assets/Editor/RenameAssets.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 08ec5285007ef964089d37440fbddca6
+guid: 4cb1b1ad9536c084caecc6e3189c00aa
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
