+++ Assets/Editor/ScriptBatch.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9a90221adc257c145a20030403b3639c
+guid: 07322183ee9d3984c99754819f6f7096
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
