+++ Assets/Samples/ExampleSample.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ad2e87374c6a1f1449d574c85dde8e3e
+guid: dd7a8db3e39d7c449b03d87809cb6a97
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
