+++ Assets/Samples/ExampleSample/ExamplePrefab.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 61badf7db01dcf5449d894b1e2fced93
+guid: be01cd1901c69d74482fb680975542a5
 PrefabImporter:
   externalObjects: {}
   userData:
