+++ Documentation/docfx.json
@@ -3,7 +3,7 @@
         {
             "src": [
                 {
-                    "src": "../Packages/com.nickmaltbie.TemplateUnityPackage",
+                    "src": "../Packages/com.nickmaltbie.RecolorShaderUnity",
                     "files": [
                         "**/*.cs"
                     ]
@@ -17,8 +17,8 @@
     ],
     "build": {
         "globalMetadata": {
-            "_appTitle": "Template Unity Package Documentation",
-            "_appFooter": "Template Unity Package Documentation",
+            "_appTitle": "RecolorShaderUnity Documentation",
+            "_appFooter": "RecolorShaderUnity Documentation",
             "_enableSearch": true
         },
         "content": [
@@ -56,7 +56,7 @@
         ],
         "overwrite": [
             {
-                "src": "../Packages/com.nickmaltbie.TemplateUnityPackage",
+                "src": "../Packages/com.nickmaltbie.RecolorShaderUnity",
                 "files": [
                     "**/*.cs"
                 ]
@@ -73,7 +73,7 @@
             }
         ],
         "sitemap": {
-            "baseUrl": "https://nickmaltbie.com/TemplateUnityPackage/docs",
+            "baseUrl": "https://nickmaltbie.com/RecolorShaderUnity/docs",
             "changefreq": "weekly",
             "fileOptions": {
                 "latest/api/*": {
