+++ Packages/com.nickmaltbie.recolorshaderunity/CHANGELOG.md.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: faf341d8dc88baa40887e55a646f8429
-AssemblyDefinitionImporter:
+guid: dde46ebd59d6f294ca7fb2ac3c938731
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
