+++ Packages/com.nickmaltbie.recolorshaderunity/LICENSE.txt.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: f8fc89548295c2546b0685bed0925b52
-folderAsset: yes
-DefaultImporter:
+guid: 73c1020fb21e07b4cad6c8a48f95da93
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
