+++ Packages/com.nickmaltbie.recolorshaderunity/README.md.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 218f47f301e83fa47a1e267bc62390f1
+guid: 84a5b317903b7f641b4cf66321209468
 TextScriptImporter:
   externalObjects: {}
   userData:
