+++ Packages/com.nickmaltbie.recolorshaderunity/Scripts.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: fbb37ea4ab6d4c94593c8f70ac111607
-TextScriptImporter:
+guid: 84b3dc3b103533847b772bc2607babca
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
