+++ Packages/com.nickmaltbie.recolorshaderunity/Scripts/SampleScript.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3af6e1d4a92888b4a80b60119dd406be
+guid: 68e244899a179704f85d6a8f049d570c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
