+++ Packages/com.nickmaltbie.recolorshaderunity/Scripts/nickmaltbie.RecolorShaderUnity.asmdef.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: c0e5fe145b1e8af4b8f3d1be0f0bfec7
-folderAsset: yes
-DefaultImporter:
+guid: 199b9e3fbc9d4d446ab7e03610f0ddab
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
