+++ Packages/com.nickmaltbie.recolorshaderunity/Tests.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: dc70346e3b16398409055b50b81491c6
-PackageManifestImporter:
+guid: 5cac0adf89934c148827cc949482539b
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
