+++ Packages/com.nickmaltbie.recolorshaderunity/Tests/EditMode.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: a7a6ad7e1c6a13149b40b7698db96e38
-TextScriptImporter:
+guid: 8486440db4c080746904e5eb790e7fa6
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
