+++ Packages/com.nickmaltbie.recolorshaderunity/Tests/EditMode/EditModeTests.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fbb06df6b50ed474ab2e8cb8c7177ebe
+guid: 3142235f5d330114b928afe8b4733a1a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
