+++ Packages/com.nickmaltbie.recolorshaderunity/Tests/EditMode/nickmaltbie.RecolorShaderUnity.Tests.EditMode.asmdef.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fc1f856a37f122e49ad7922398a4541a
+guid: eba782018885f124ab2574f2c5678097
 AssemblyDefinitionImporter:
   externalObjects: {}
   userData:
