+++ Packages/com.nickmaltbie.recolorshaderunity/Tests/PlayMode.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f9165928ac427214d886d8162892680f
+guid: aaf7346a1aebd254ba04b5267a137599
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
