+++ Packages/com.nickmaltbie.recolorshaderunity/Tests/PlayMode/PlayModeTests.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b64f585020a59f74ba9dc1c718671a8c
+guid: 7bbca9d3fd60cbf4a86db1d34383c8fd
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
