+++ Packages/com.nickmaltbie.recolorshaderunity/Tests/PlayMode/nickmaltbie.RecolorShaderUnity.Tests.PlayMode.asmdef
@@ -1,15 +1,13 @@
 {
-    "name": "nickmaltbie.TemplateUnityPackage.Tests.EditMode",
-    "rootNamespace": "nickmaltbie.TemplateUnityPackage.Tests.EditMode",
+    "name": "nickmaltbie.RecolorShaderUnity.Tests.PlayMode",
+    "rootNamespace": "nickmaltbie.RecolorShaderUnity.Tests.PlayMode",
     "references": [
-        "nickmaltbie.TemplateUnityPackage"
+        "nickmaltbie.RecolorShaderUnity"
     ],
     "optionalUnityReferences": [
         "TestAssemblies"
     ],
-    "includePlatforms": [
-        "Editor"
-    ],
+    "includePlatforms": [],
     "excludePlatforms": [],
     "allowUnsafeCode": false,
     "overrideReferences": true,
