+++ Packages/com.nickmaltbie.recolorshaderunity/Tests/PlayMode/nickmaltbie.RecolorShaderUnity.Tests.PlayMode.asmdef.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 0a260c49d7d6d8046a4b7a23f5460063
-folderAsset: yes
-DefaultImporter:
+guid: 46671d22de5699642ba9bcf2507ca697
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
