+++ Packages/com.nickmaltbie.recolorshaderunity/package.json.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 9cc0d73a64630ea4094edf2aa62286e2
-AssemblyDefinitionImporter:
+guid: 22431f85e0fac2f4f9e712d2db83f6ca
+PackageManifestImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
