+++ README.md
@@ -1,6 +1,6 @@
-# Template Unity Package
+# RecolorShaderUnity
 
-Nick Maltbie's Template Unity Package for easily creating and sharing unity
+Nick Maltbie's RecolorShaderUnity for easily creating and sharing unity
 packages over git and NPM.
 
 ![Example importing package from git.](Demo/project_demo.png)
@@ -8,8 +8,8 @@ packages over git and NPM.
 If you want to host and share code between unity projects easily,
 use this project to configure unity builds.
 
-Template Unity Package is an open source project hosted at
-[https://github.com/nicholas-maltbie/TemplateUnityPackage](https://github.com/nicholas-maltbie/TemplateUnityPackage)
+RecolorShaderUnity is an open source project hosted at
+[https://github.com/nicholas-maltbie/RecolorShaderUnity](https://github.com/nicholas-maltbie/RecolorShaderUnity)
 
 This is an open source project licensed under a [MIT License](LICENSE.txt).
 
@@ -22,7 +22,7 @@ Install this version of Unity from Unity Hub using this link:
 ## Template Components
 
 Parts of template project, please see the documentation at
-[Template Unity Package Introduction](https://nickmaltbie.com/TemplateUnityPackage/docs/manual/introduction.html)
+[RecolorShaderUnity Introduction](https://nickmaltbie.com/RecolorShaderUnity/docs/manual/introduction.html)
 for more details on how to use the project.
 
 1. Automated script to rename all scripts in project.
@@ -34,13 +34,13 @@ for more details on how to use the project.
 ## Documentation
 
 Documentation on the project and scripting API is found at
-[https://nickmaltbie.com/TemplateUnityPackage/docs/](https://nickmaltbie.com/TemplateUnityPackage/docs/)
+[https://nickmaltbie.com/RecolorShaderUnity/docs/](https://nickmaltbie.com/RecolorShaderUnity/docs/)
 for the latest version of the codebase.
 
 ## Demo
 
 You can see a demo of the project running here:
-[https://nickmaltbie.com/TemplateUnityPackage/](https://nickmaltbie.com/TemplateUnityPackage/).
+[https://nickmaltbie.com/RecolorShaderUnity/](https://nickmaltbie.com/RecolorShaderUnity/).
 The project hosted on the website is up to date with the most recent
 version on the `main` branch of this github repo
 and is automatically deployed with each update to the codebase.
@@ -49,4 +49,4 @@ and is automatically deployed with each update to the codebase.
 
 The samples in the project include:
 
-* ExampleSample - Example sample for Template Unity Package.
+* ExampleSample - Example sample for RecolorShaderUnity.
